Use defaultConfig with defaultDecoder
While the first scraping target didn't return the `Content-Type` header
and returned UTF-8, the two other targets do return that header and
return Latin-1. Hence, we just rely on the default config and the first
scraping target's results include some gibberish. This is fine since we
only use the results for comparison anyway; they are never displayed to
the user.